requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.0
openai==2.3.0
outcome==1.3.0.post0
packaging==25.0
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
pyperclip==1.11.0
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.5
selenium==4.36.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.67.1
trio==0.31.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.2.0
